+++ itachallenge-challenge/src/main/java/com/itachallenge/gamification/service/UserScoreServiceImpl.java
@@ -57,8 +57,8 @@ public class UserScoreServiceImpl implements UserScoreService {
             return Mono.error(new IllegalArgumentException("ActivityType cannot be null"));
         }
 
-        if (username == null || username.isBlank()) {
-            return Mono.error(new IllegalArgumentException("username cannot be null or blank"));
+        if (userId == null){
+            return Mono.error(new IllegalArgumentException("userId cannot be null or blank"));
         }
 
         if (type == ActivityType.CHALLENGE_COMPLETED && challengeId == null) {